合并改动

+++ Scripts/AutoBuild_Function.sh
@@ -7,6 +7,7 @@ GET_TARGET_INFO() {
 	Diy_Core
 	Home="${GITHUB_WORKSPACE}/openwrt"
 	[ -f "${GITHUB_WORKSPACE}/Openwrt.info" ] && . ${GITHUB_WORKSPACE}/Openwrt.info
+	[[ "${Short_Firmware_Date}" == true ]] && Compile_Date="$(echo ${Compile_Date} | cut -c1-8)"
 	Owner_Repo="$(grep "https://github.com/[a-zA-Z0-9]" ${GITHUB_WORKSPACE}/.git/config | cut -c8-100)"
 	Source_Repo="$(grep "https://github.com/[a-zA-Z0-9]" ${Home}/.git/config | cut -c8-100)"
 	Source_Owner="$(echo "${Source_Repo}" | egrep -o "[a-z]+" | awk 'NR==4')"
@@ -37,7 +38,7 @@ GET_TARGET_INFO() {
 		x86_Test="$(egrep -o "CONFIG_TARGET.*DEVICE.*=y" .config | sed -r 's/CONFIG_TARGET_(.*)_DEVICE_(.*)=y/\1/')"
 		[[ -n "${x86_Test}" ]] && break
 		x86_Test="$(egrep -o "CONFIG_TARGET.*Generic=y" .config | sed -r 's/CONFIG_TARGET_(.*)_Generic=y/\1/')"
-		[[ -z "${x86_Test}" ]] && TIME "Can not obtain the TARGET_PROFILE !" && exit 1
+		[[ -z "${x86_Test}" ]] && TIME "[ERROR] Can not obtain the TARGET_PROFILE !" && exit 1
 	done
 	[[ "${x86_Test}" == "x86_64" ]] && {
 		TARGET_PROFILE="x86_64"
@@ -45,21 +46,24 @@ GET_TARGET_INFO() {
 		TARGET_PROFILE="$(egrep -o "CONFIG_TARGET.*DEVICE.*=y" .config | sed -r 's/.*DEVICE_(.*)=y/\1/')"
 	}
 	[[ -z "${TARGET_PROFILE}" ]] && TARGET_PROFILE="${Default_Device}"
-	case ${TARGET_PROFILE} in
-	x86_64)
+	[[ "${TARGET_PROFILE}" == x86_64 ]] && {
 		[[ "$(cat ${Home}/.config)" =~ "CONFIG_TARGET_IMAGES_GZIP=y" ]] && {
 			Firmware_Type=img.gz
 		} || {
 			Firmware_Type=img
 		}
-	;;
-	*)
+	}
+	TARGET_BOARD="$(awk -F '[="]+' '/TARGET_BOARD/{print $2}' .config)"
+	case ${TARGET_BOARD} in
+	ramips | reltek | ipq40xx | ath79)
 		Firmware_Type=bin
 	;;
+	rockchip)
+		Firmware_Type=img.gz
+	;;
 	esac
-	TARGET_BOARD="$(awk -F '[="]+' '/TARGET_BOARD/{print $2}' .config)"
 	TARGET_SUBTARGET="$(awk -F '[="]+' '/TARGET_SUBTARGET/{print $2}' .config)"
-	
+
 	echo "Firmware_Type=${Firmware_Type}" > ${Home}/TARGET_INFO
 	echo "TARGET_PROFILE=${TARGET_PROFILE}" >> ${Home}/TARGET_INFO
 	echo "Openwrt_Version=${Openwrt_Version}" >> ${Home}/TARGET_INFO
@@ -68,12 +72,12 @@ GET_TARGET_INFO() {
 	echo "TARGET_SUBTARGET=${TARGET_SUBTARGET}" >> ${Home}/TARGET_INFO
 	echo "Home=${Home}" >> ${Home}/TARGET_INFO
 	echo "Current_Branch=${Current_Branch}" >> ${Home}/TARGET_INFO
-	
+
 	echo "${Openwrt_Version}" > ${AB_Firmware_Info}
 	echo "${Owner_Repo}" >> ${AB_Firmware_Info}
 	echo "${TARGET_PROFILE}" >> ${AB_Firmware_Info}
 	echo "${Firmware_Type}" >> ${AB_Firmware_Info}
-	
+
 	echo "Author: ${Author}"
 	echo "Author Github: ${Owner_Repo}"
 	echo "Firmware Version: ${Openwrt_Version}"
@@ -81,16 +85,16 @@ GET_TARGET_INFO() {
 	echo "Source: ${Source_Repo}"
 	echo "Source Author: ${Source_Owner}"
 	echo "Source Branch: ${Current_Branch}"
-	echo "TARGET_PTOFILE: ${TARGET_PROFILE}"
+	echo "TARGET_PROFILE: ${TARGET_PROFILE}"
 	echo "TARGET_BOARD: ${TARGET_BOARD}"
 	echo "TARGET_SUBTARGET: ${TARGET_SUBTARGET}"
-	
+
 	TIME "[Preload Info] All done !"
 }
 
 Firmware-Diy_Base() {
 	GET_TARGET_INFO
-	Auto_ExtraPackages
+	Auto_AddPackage
 	chmod +x -R ${GITHUB_WORKSPACE}/Scripts
 	chmod +x -R ${GITHUB_WORKSPACE}/CustomFiles
 	chmod +x -R ${GITHUB_WORKSPACE}/CustomPackages
@@ -98,24 +102,24 @@ Firmware-Diy_Base() {
 		Replace_File Scripts/AutoBuild_Tools.sh package/base-files/files/bin
 	}
 	[[ "${INCLUDE_AutoUpdate}" == true ]] && {
-		ExtraPackages git lean luci-app-autoupdate https://github.com/Hyy2001X main
+		AddPackage git lean luci-app-autoupdate https://github.com/Hyy2001X main
 		Replace_File Scripts/AutoUpdate.sh package/base-files/files/bin
 	}
 	[[ "${INCLUDE_Theme_Argon}" == true ]] && {
 		case ${Source_Owner} in
 		coolsnowwolf)
-				ExtraPackages git lean luci-theme-argon https://github.com/jerrykuku 18.06
+			AddPackage git lean luci-theme-argon https://github.com/jerrykuku 18.06
 		;;
 		*)
 			case ${Current_Branch} in
 			19.07)
-				ExtraPackages git other luci-theme-argon https://github.com/jerrykuku v2.2.5
+				AddPackage git other luci-theme-argon https://github.com/jerrykuku v2.2.5
 			;;
 			21.02)
-				ExtraPackages git other luci-theme-argon https://github.com/jerrykuku
+				AddPackage git other luci-theme-argon https://github.com/jerrykuku
 			;;
 			18.06)
-				ExtraPackages git other luci-theme-argon https://github.com/jerrykuku 18.06
+				AddPackage git other luci-theme-argon https://github.com/jerrykuku 18.06
 			;;
 			*)
 				TIME "[ERROR] Unknown source branch: [${Current_Branch}] !"
@@ -123,32 +127,42 @@ Firmware-Diy_Base() {
 			esac
 		;;
 		esac
+	}	
+	[[ -n "${Default_IP_Address}" ]] && {
+		if [[ "${Default_IP_Address}" =~ ^[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}$ ]];then
+			Old_IP_Address=$(awk -F '[="]+' '/ipaddr:-/{print $3}' package/base-files/files/bin/config_generate | awk 'NR==1')
+			if [[ ! "${Default_IP_Address}" == "${Old_IP_Address}" ]];then
+				TIME "Setting default IP Address to ${Default_IP_Address} ..."
+				sed -i "s/${Old_IP_Address}/${Default_IP_Address}/g" package/base-files/files/bin/config_generate
+			fi
+		else
+			TIME "[ERROR] ${Default_IP_Address} is not an IP Address !"
+		fi
 	}
 	[ -f package/base-files/files/bin/AutoUpdate.sh ] && {
 		AutoUpdate_Version=$(awk 'NR==6' package/base-files/files/bin/AutoUpdate.sh | awk -F '[="]+' '/Version/{print $2}')
 	} || AutoUpdate_Version=OFF
-
 	Replace_File CustomFiles/Depends/profile package/base-files/files/etc
+	sed -i '/profile/d' package/base-files/files/lib/upgrade/keep.d/base-files-essential
 	case ${Source_Owner} in
 	coolsnowwolf)
 		Replace_File CustomFiles/Depends/coremark_lede.sh package/lean/coremark coremark.sh
 		Replace_File CustomFiles/Depends/cpuinfo_x86 package/lean/autocore/files/x86/sbin cpuinfo
-		ExtraPackages git lean helloworld https://github.com/fw876 master
+		AddPackage git other helloworld https://github.com/fw876 master
 		sed -i 's/143/143,8080/' $(PKG_Finder d package luci-app-ssr-plus)/root/etc/init.d/shadowsocksr
-		sed -i "s?iptables?#iptables?g" ${Version_File} > /dev/null 2>&1
-		sed -i "s?${Old_Version}?${Old_Version} Compiled by ${Author} [${Display_Date}]?g" $Version_File
+		sed -i "s?iptables?#iptables?g" ${Version_File}
+		sed -i "s?${Old_Version}?${Old_Version} Compiled by ${Author} [${Display_Date}]?g" ${Version_File}
 		[[ "${INCLUDE_DRM_I915}" == true ]] && Replace_File CustomFiles/Depends/i915-5.4 target/linux/x86 config-5.4
 	;;
 	immortalwrt)
-		sed -i 's/143/143,8080/' package/lean/luci-app-ssr-plus/root/etc/init.d/shadowsocksr
+		sed -i 's/143/143,8080/' $(PKG_Finder d package luci-app-ssr-plus)/root/etc/init.d/shadowsocksr
 		Replace_File CustomFiles/Depends/coremark_ImmortalWrt.sh package/base-files/files/etc coremark.sh
 		Replace_File CustomFiles/Depends/ImmortalWrt package/base-files/files/etc openwrt_release
 		Replace_File CustomFiles/Depends/cpuinfo_x86 package/lean/autocore/files/x86/sbin cpuinfo
-		sed -i "s?Template?Compiled by ${Author} [${Display_Date}]?g" $Version_File
+		sed -i "s?Template?Compiled by ${Author} [${Display_Date}]?g" ${Version_File}
 		[[ "${INCLUDE_DRM_I915}" == true ]] && Replace_File CustomFiles/Depends/i915-4.19 target/linux/x86 config-4.19
 	;;
 	esac
-	
 	case ${Source_Owner} in
 	immortalwrt)
 		Replace_File CustomFiles/Depends/banner package/lean/default-settings/files openwrt_banner
@@ -161,7 +175,6 @@ Firmware-Diy_Base() {
 		sed -i "s?Openwrt?Openwrt ${Openwrt_Version} / AutoUpdate ${AutoUpdate_Version}?g" package/base-files/files/etc/banner
 	;;
 	esac
-
 	[[ "${INCLUDE_Obsolete_PKG_Compatible}" == true ]] && {
 		TIME "Start to run Obsolete_Package_Compatible Scripts ..."
 		[[ ${Source_Owner} == openwrt ]] && {
@@ -169,8 +182,7 @@ Firmware-Diy_Base() {
 			19.07 | 21.02)
 				Replace_File CustomFiles/Patches/0003-upx-ucl-${Current_Branch}.patch ./
 				cat 0003-upx-ucl-${Current_Branch}.patch | patch -p1 > /dev/null 2>&1
-				ExtraPackages svn ../feeds/packages/lang golang https://github.com/coolsnowwolf/packages/trunk/lang
-		
+				AddPackage svn ../feeds/packages/lang golang https://github.com/coolsnowwolf/packages/trunk/lang
 				TIME "Start to convert zh-cn translation files to zh_Hans ..."
 				Replace_File Scripts/Convert_Translation.sh package
 				cd ./package
@@ -190,7 +202,6 @@ Firmware-Diy_Base() {
 
 PS_Firmware() {
 	. TARGET_INFO
-	ls bin/targets/${TARGET_BOARD}/${TARGET_SUBTARGET}
 	case ${Source_Owner} in
 	immortalwrt)
 		_Firmware=immortalwrt
@@ -222,32 +233,32 @@ PS_Firmware() {
 	case "${TARGET_PROFILE}" in
 	x86_64)
 		cd ${Firmware_Path}
-		Legacy_Firmware=${_Firmware}-${TARGET_BOARD}-${TARGET_SUBTARGET}-${_Legacy_Firmware}.${Firmware_Type}
-		EFI_Firmware=${_Firmware}-${TARGET_BOARD}-${TARGET_SUBTARGET}-${_EFI_Firmware}.${Firmware_Type}
+		Legacy_Firmware="${_Firmware}-${TARGET_BOARD}-${TARGET_SUBTARGET}-${_Legacy_Firmware}.${Firmware_Type}"
+		EFI_Firmware="${_Firmware}-${TARGET_BOARD}-${TARGET_SUBTARGET}-${_EFI_Firmware}.${Firmware_Type}"
 		AutoBuild_Firmware="AutoBuild-${TARGET_PROFILE}-${Openwrt_Version}"
 		echo "[Preload Info] Legacy_Firmware: ${Legacy_Firmware}"
 		echo "[Preload Info] UEFI_Firmware: ${EFI_Firmware}"
 		echo "[Preload Info] AutoBuild_Firmware: ${AutoBuild_Firmware}"
-		[ -f "${Legacy_Firmware}" ] && {
+		if [ -f "${Legacy_Firmware}" ];then
 			_MD5=$(md5sum ${Legacy_Firmware} | cut -d ' ' -f1)
 			_SHA256=$(sha256sum ${Legacy_Firmware} | cut -d ' ' -f1)
 			touch ${Home}/bin/Firmware/${AutoBuild_Firmware}.detail
 			echo -e "\nMD5:${_MD5}\nSHA256:${_SHA256}" > ${Home}/bin/Firmware/${AutoBuild_Firmware}-Legacy.detail
 			mv -f ${Legacy_Firmware} ${Home}/bin/Firmware/${AutoBuild_Firmware}-Legacy.${Firmware_Type}
 			TIME "Legacy Firmware is detected !"
-		} || {
+		else
 			TIME "[ERROR] Legacy Firmware is not detected !"
-		}
-		[ -f "${EFI_Firmware}" ] && {
+		fi
+		if [ -f "${EFI_Firmware}" ];then
 			_MD5=$(md5sum ${EFI_Firmware} | cut -d ' ' -f1)
 			_SHA256=$(sha256sum ${EFI_Firmware} | cut -d ' ' -f1)
 			touch ${Home}/bin/Firmware/${AutoBuild_Firmware}-UEFI.detail
 			echo -e "\nMD5:${_MD5}\nSHA256:${_SHA256}" > ${Home}/bin/Firmware/${AutoBuild_Firmware}-UEFI.detail
 			cp ${EFI_Firmware} ${Home}/bin/Firmware/${AutoBuild_Firmware}-UEFI.${Firmware_Type}
 			TIME "UEFI Firmware is detected !"
-		} || {
+		else
 			TIME "[ERROR] UEFI Firmware is not detected !"
-		}
+		fi
 	;;
 	*)
 		cd ${Home}
@@ -256,15 +267,15 @@ PS_Firmware() {
 		AutoBuild_Detail="AutoBuild-${TARGET_PROFILE}-${Openwrt_Version}.detail"
 		echo "[Preload Info] Default_Firmware: ${Default_Firmware}"
 		echo "[Preload Info] AutoBuild_Firmware: ${AutoBuild_Firmware}"
-		[ -f ${Firmware_Path}/${Default_Firmware} ] && {
+		if [ -f "${Firmware_Path}/${Default_Firmware}" ];then
 			mv -f ${Firmware_Path}/${Default_Firmware} bin/Firmware/${AutoBuild_Firmware}
 			_MD5=$(md5sum bin/Firmware/${AutoBuild_Firmware} | cut -d ' ' -f1)
 			_SHA256=$(sha256sum bin/Firmware/${AutoBuild_Firmware} | cut -d ' ' -f1)
 			echo -e "\nMD5:${_MD5}\nSHA256:${_SHA256}" > bin/Firmware/${AutoBuild_Detail}
 			TIME "Firmware is detected !"
-		} || {
+		else
 			TIME "[ERROR] Firmware is not detected !"
-		}
+		fi
 	;;
 	esac
 	cd ${Home}
@@ -303,28 +314,32 @@ PKG_Finder() {
 	unset _PKG_TYPE _PKG_DIR _PKG_NAME
 }
 
-Auto_ExtraPackages() {
+Auto_AddPackage() {
 	COMMON_FILE="${GITHUB_WORKSPACE}/CustomPackages/Common"
 	TARGET_FILE="${GITHUB_WORKSPACE}/CustomPackages/${TARGET_PROFILE}"
-	Auto_ExtraPackages_mod ${COMMON_FILE}
-	Auto_ExtraPackages_mod ${TARGET_FILE}
+	Auto_AddPackage_mod ${COMMON_FILE}
+	Auto_AddPackage_mod ${TARGET_FILE}
 }
 
-Auto_ExtraPackages_mod() {
-    _FILENAME=${1}
+Auto_AddPackage_mod() {
+	[[ $# != 1 ]] && {
+		TIME "[ERROR] Error options: [$#] [$*] !"
+		return 0
+	}
+	_FILENAME=${1}
 	echo "" >> ${_FILENAME}
-    [ -f "${_FILENAME}" ] && {
-	    TIME "Loading Custom Packages list: [${_FILENAME}]..."
-	    cat ${_FILENAME} | sed '/^$/d' | while read X
-	    do
-	    	[[ "${X}" != "" ]] && [[ -n ${X} ]] && ExtraPackages ${X}
-	    	unset X
-	    done
+	[ -f "${_FILENAME}" ] && {
+		TIME "Loading Custom Packages list: [${_FILENAME}]..."
+		cat ${_FILENAME} | sed '/^$/d' | while read X
+		do
+			[[ "${X}" != "" ]] && [[ -n ${X} ]] && AddPackage ${X}
+			unset X
+		done
 	}
 	unset _FILENAME
 }
 
-ExtraPackages() {
+AddPackage() {
 	[[ $# -lt 4 ]] && {
 		TIME "[ERROR] Error options: [$#] [$*] !"
 		return 0
@@ -332,20 +347,18 @@ ExtraPackages() {
 	case ${1} in
 	git | svn)
 		PKG_PROTO=${1}
+		PKG_DIR=${2}
+		PKG_NAME=${3}
+		REPO_URL=${4}
+		REPO_BRANCH=${5}
 	;;
 	*)
 		return 0
 	;;
 	esac
-	
-	PKG_DIR=${2}
-	PKG_NAME=${3}
-	REPO_URL=${4}
-	REPO_BRANCH=${5}
-
-	Mkdir "package/${PKG_DIR}"
+	Mkdir package/${PKG_DIR}
 	[ -d "package/${PKG_DIR}/${PKG_NAME}" ] && {
-		TIME "Removing old package [${PKG_NAME}] ..."
+		TIME "Removing old package: [${PKG_NAME}] ..."
 		rm -rf "package/${PKG_DIR}/${PKG_NAME}"
 	}
 	TIME "Checking out package [${PKG_NAME}] to package/${PKG_DIR} ..."
@@ -359,7 +372,7 @@ ExtraPackages() {
 		svn checkout ${REPO_URL}/${PKG_NAME} ${PKG_NAME} > /dev/null 2>&1
 	;;
 	esac
-	[ -f ${PKG_NAME}/Makefile ] || [ -f ${PKG_NAME}/README* ] && {
+	[ -f ${PKG_NAME}/Makefile ] || [ -f ${PKG_NAME}/README* ] || [[ -n "$(ls -A ${PKG_NAME})" ]] && {
 		mv -f "${PKG_NAME}" "package/${PKG_DIR}"
 	} || {
 		TIME "[ERROR] Package [${PKG_NAME}] is not detected!"
@@ -375,7 +388,7 @@ Replace_File() {
 	FILE_NAME=${1}
 	PATCH_DIR=${GITHUB_WORKSPACE}/openwrt/${2}
 	FILE_RENAME=${3}
-	
+
 	Mkdir ${PATCH_DIR}
 	[ -f "${GITHUB_WORKSPACE}/${FILE_NAME}" ] && _TYPE1="f" && _TYPE2="File"
 	[ -d "${GITHUB_WORKSPACE}/${FILE_NAME}" ] && _TYPE1="d" && _TYPE2="Folder"
